refactor(web): Phase 7 — lift terminals into 3-tier (Infra → Service → API)
Implements GitHub issue #318. Migrates the terminal domain into the
3-tier architecture described in `docs/web-architecture.md`, alongside
the Phase 0/1/3 work that already landed.

Layout:

- **Infra:** `server/infra/terminals/terminal-pool.ts` — owns the PTY
  lifecycle (`node-pty` spawn/kill, scrollback buffering, output
  listeners, reverse workspace→terminal index). Singleton `terminalPool`.
- **Service:** `server/services/terminal-service.ts` — coordinates the
  pool with workspace resolution, dockview layout persistence, and the
  workspace status event bus. Absorbs the `workspace.terminal` config
  parsing that used to live in `lib/terminal-config.ts`.
- **API:** `server/api/terminals/router.ts` — thin tRPC sub-router for
  the `terminal` and `terminalLayout` keys. Composite export
  `terminalsRouters` is merged into the root router by
  `server/api/router.ts` so the wire shape is preserved.
- **WS:** `server/api/terminals/ws.ts` — terminal WebSocket handler,
  moved from `lib/terminal-ws.ts`. Routes through `terminalService`.

Cleanup:

- Removes the inline `terminal`/`terminalLayout` sub-routers from
  the legacy `trpc/router.ts` (now merged in via `mergeRouters`).
- Deletes `lib/terminal-manager.ts`, `lib/terminal-ws.ts`,
  `lib/terminal-config.ts`.
- `start-server.ts` shutdown path calls `terminalService.killAll()`.

The `lib/terminal-layout-manager.ts` and `lib/terminal-selection.ts`
modules stay in place for now — the layout manager is consumed via thin
service pass-throughs (see the inline comment in `terminal-service.ts`)
and the selection helpers are frontend-only.

Intentional behaviour widening: the WebSocket client-initiated
`{ type: "close" }` message now routes through `terminalService.kill()`,
which (matching the long-standing `terminal.kill` tRPC mutation) emits a
`terminal-killed` workspace event and removes the panel from the saved
dockview layout in addition to killing the PTY. The legacy WS handler
only killed the PTY, leaving an orphan layout entry that would fail
silently on reload — the unified path is the correct shape.

Closes #318
