+++ apps/web/src/server/services/workspace-service.ts
@@ -30,10 +30,9 @@ import {
 // capturing `submitTask` (or anything else here) at module load.
 import { submitTask } from "../../lib/task-runner";
 import { deleteWorkspaceTasks } from "../../lib/task-store";
-import { deleteTerminalLayout } from "../../lib/terminal-layout-manager";
-import { killWorkspaceTerminals } from "../../lib/terminal-manager";
 import { emit } from "../../lib/watcher";
 import { WorkspaceQueries } from "../infra/db/queries/workspaces";
+import { terminalService } from "./terminal-service";
 // FRAGILE: ESM cycle leg #2 — `./cronjob-service` imports `submitTask`
 // from `lib/task-runner`, which imports `lib/workspace`, which imports
 // `workspaceService` from this file. Same live-binding constraint as the
@@ -173,10 +172,10 @@ export class PlainProjectError extends Error {
  *     service uses them directly today; the router-facing contract is
  *     unchanged.
  *   - **Workspace-scoped side-effect cleanup.** `removeWorkspaceChats`,
- *     `killWorkspaceTerminals`, `cronjobService.removeForKey`, etc. live
- *     in their own domain modules. Each will migrate to its own service
- *     in a later phase; the orchestration is centralized here for now so
- *     the remove flow remains atomic from the router's perspective.
+ *     `terminalService.killWorkspace`, `cronjobService.removeForKey`, etc.
+ *     live in their own domain modules. Each will migrate to its own
+ *     service in a later phase; the orchestration is centralized here for
+ *     now so the remove flow remains atomic from the router's perspective.
  *
  * Stateless aside from its `queries` dependency, so a single shared
  * instance is safe across callers.
@@ -403,8 +402,8 @@ export class WorkspaceService {
     deleteBrowserLayout(workspaceId);
 
     // Kill any running terminal PTY sessions + layout
-    killWorkspaceTerminals(workspaceId);
-    deleteTerminalLayout(workspaceId);
+    terminalService.killWorkspace(workspaceId);
+    terminalService.deleteLayout(workspaceId);
 
     // Kill any running language server processes
     killWorkspaceServers(workspaceId);
