+++ apps/web/src/trpc/router.ts
@@ -1,5 +1,4 @@
 import { execFile, spawn } from "node:child_process";
-import { randomUUID } from "node:crypto";
 import { existsSync, constants as fsConstants, realpathSync, statSync } from "node:fs";
 import { cp, mkdir, open, readdir, readFile, rename, rm, stat, writeFile } from "node:fs/promises";
 import { dirname, extname, isAbsolute, join, resolve, sep } from "node:path";
@@ -91,26 +90,12 @@ import {
 } from "../lib/state";
 import { abortTask, cancelTask, getTask, submitTask, TaskConflictError } from "../lib/task-runner";
 import { listTasks, loadTask } from "../lib/task-store";
-import { loadWorkspaceTerminalConfig } from "../lib/terminal-config";
-import {
-  getTerminalLayout,
-  removeTerminalFromLayout,
-  saveTerminalLayout,
-} from "../lib/terminal-layout-manager";
-import {
-  getScrollback,
-  getTerminalSession,
-  killTerminal,
-  listTerminals,
-  spawnTerminal,
-  subscribeTerminalOutput,
-  writeToTerminal,
-} from "../lib/terminal-manager";
 import { getTunnelStatus, startTunnel, stopTunnel } from "../lib/tunnel";
 import { saveUploadedFiles, saveUploadedFilesDetailed } from "../lib/upload-utils";
 import { emit, subscribe as subscribeStatus } from "../lib/watcher";
 import { resolveWorkspace } from "../lib/workspace";
 import { publicProcedure, t } from "../server/api/trpc";
+import { terminalService } from "../server/services/terminal-service";
 
 const log = createLogger("trpc");
 
@@ -358,10 +343,7 @@ const workspaceRouter = t.router({
   getTerminalConfig: publicProcedure
     .input(z.object({ workspaceId: z.string() }))
     .query(({ input }) => {
-      const workspace = resolveWorkspace(input.workspaceId);
-      if (!workspace) return { config: null };
-      const config = loadWorkspaceTerminalConfig(workspace.worktree.path, workspace.project.path);
-      return { config };
+      return { config: terminalService.getWorkspaceConfig(input.workspaceId) };
     }),
 
   /**
@@ -3571,151 +3553,10 @@ const queueRouter = t.router({
     }),
 });
 
-// ---------------------------------------------------------------------------
-// Terminal Layout (split pane tree persistence)
-// ---------------------------------------------------------------------------
-
-const terminalLayoutRouter = t.router({
-  get: publicProcedure.input(z.object({ workspaceId: z.string() })).query(({ input }) => {
-    return { tree: getTerminalLayout(input.workspaceId) };
-  }),
-
-  save: publicProcedure
-    .input(z.object({ workspaceId: z.string(), tree: z.unknown() }))
-    .mutation(({ input }) => {
-      saveTerminalLayout(input.workspaceId, input.tree);
-      return { ok: true };
-    }),
-});
-
-// ---------------------------------------------------------------------------
-// Terminal
-// ---------------------------------------------------------------------------
-
-const terminalRouter = t.router({
-  list: publicProcedure.input(z.object({ workspaceId: z.string() })).query(({ input }) => {
-    return { terminals: listTerminals(input.workspaceId) };
-  }),
-
-  create: publicProcedure
-    .input(
-      z.object({
-        workspaceId: z.string(),
-        id: z.string().optional(),
-        command: z.string().optional(),
-        cwd: z.string().optional(),
-        env: z.record(z.string()).optional(),
-      }),
-    )
-    .mutation(async ({ input }) => {
-      const terminalId = input.id ?? randomUUID();
-      // `spawnTerminal` registers the session and writes to the saved
-      // dockview layout (mirrors `createChat` / `createBrowser`); no
-      // separate `addTerminalToLayout` call needed here.
-      const session = await spawnTerminal(input.workspaceId, terminalId, {
-        command: input.command,
-        cwd: input.cwd,
-        env: input.env,
-      });
-      emit({ kind: "terminal-created", workspaceId: input.workspaceId, terminalId });
-      return { terminalId, workspaceId: input.workspaceId, pid: session.pty.pid };
-    }),
-
-  send: publicProcedure
-    .input(z.object({ terminalId: z.string(), data: z.string() }))
-    .mutation(({ input }) => {
-      const ok = writeToTerminal(input.terminalId, input.data);
-      if (!ok) {
-        throw new TRPCError({
-          code: "NOT_FOUND",
-          message: `Terminal not found: ${input.terminalId}`,
-        });
-      }
-      return { ok: true };
-    }),
-
-  output: publicProcedure
-    .input(z.object({ terminalId: z.string(), lines: z.number().int().positive().optional() }))
-    .query(({ input }) => {
-      const output = getScrollback(input.terminalId, input.lines ?? undefined);
-      if (output == null) {
-        throw new TRPCError({
-          code: "NOT_FOUND",
-          message: `Terminal not found: ${input.terminalId}`,
-        });
-      }
-      return { output };
-    }),
-
-  kill: publicProcedure.input(z.object({ terminalId: z.string() })).mutation(({ input }) => {
-    const session = getTerminalSession(input.terminalId);
-    const workspaceId = session?.workspaceId;
-    killTerminal(input.terminalId);
-    if (workspaceId) {
-      removeTerminalFromLayout(workspaceId, input.terminalId);
-      emit({ kind: "terminal-killed", workspaceId, terminalId: input.terminalId });
-    }
-    return { ok: true };
-  }),
-
-  stream: publicProcedure
-    .input(
-      z.object({
-        terminalId: z.string(),
-        replay: z.boolean().optional().default(true),
-      }),
-    )
-    .subscription(async function* (opts) {
-      const { terminalId, replay } = opts.input;
-
-      // Check if terminal exists
-      const session = getTerminalSession(terminalId);
-      if (!session) {
-        yield { type: "error" as const, data: `Terminal not found: ${terminalId}` };
-        return;
-      }
-
-      // Replay buffered scrollback first
-      if (replay && session.scrollback.length > 0) {
-        yield { type: "output" as const, data: session.scrollback };
-      }
-
-      // Stream live output
-      const queue: string[] = [];
-      let resolve: (() => void) | null = null;
-
-      const unsubscribe = subscribeTerminalOutput(terminalId, (data: string) => {
-        queue.push(data);
-        resolve?.();
-      });
-
-      opts.signal?.addEventListener("abort", () => {
-        unsubscribe();
-        resolve?.();
-      });
-
-      try {
-        while (!opts.signal?.aborted) {
-          while (queue.length > 0) {
-            yield { type: "output" as const, data: queue.shift()! };
-          }
-
-          // Check if terminal is still alive
-          if (!getTerminalSession(terminalId)) {
-            yield { type: "exit" as const };
-            return;
-          }
-
-          await new Promise<void>((r) => {
-            resolve = r;
-          });
-          resolve = null;
-        }
-      } finally {
-        unsubscribe();
-      }
-    }),
-});
+// NOTE: the `terminal` and `terminalLayout` sub-routers were lifted into
+// `server/api/terminals/router.ts` as part of Phase 7 (issue #318). The wire
+// shape is preserved by `server/api/router.ts` via `mergeRouters` — see the
+// invariant comment there.
 
 // ---------------------------------------------------------------------------
 // App Router
@@ -3760,8 +3601,6 @@ export const appRouter = t.router({
   modes: modesRouter,
   models: modelsRouter,
   queue: queueRouter,
-  terminal: terminalRouter,
-  terminalLayout: terminalLayoutRouter,
 });
 
 // NOTE: the canonical `AppRouter` type now lives in
